Refactor: 상수 이름 신청을 대상 상태는 승인(APPROVAL), 게임 상태는 확정(CONFIRMED) 로 통일
